Update base integration test classes
